third_party/WebKit/LayoutTests/http/tests/security/cross-origin-access-over-property-descriptor-expected.txt

Cross-origin access to 'window.location.pathname' over 'get' and 'set' in property descriptor should throw a SecurityError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL pathnameDescriptor should be non-null. Was undefined
FAIL typeof pathnameDescriptor.get should be function. Threw exception TypeError: Cannot read property 'get' of undefined
FAIL typeof pathnameDescriptor.set should be function. Threw exception TypeError: Cannot read property 'set' of undefined
FAIL pathnameDescriptor.get.call(targetWindow.location) should throw SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.. Threw exception TypeError: Cannot read property 'get' of undefined.
FAIL pathnameDescriptor.set.call(targetWindow.location, 1) should throw SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.. Threw exception TypeError: Cannot read property 'set' of undefined.
PASS successfullyParsed is true

TEST COMPLETE
